+++ HTS_TEST/HTS_V400_Dispatcher.hpp
@@ -228,7 +228,7 @@ class HTS_V400_Dispatcher {
     /// 별도.
     ///       PC 벤치 `HTS_Fractal_Channel_Compare` 가 IR 경로 지연 proxy에
     ///       동기화.
-    static constexpr double IR_HARQ_RTT_MS = 4.0;
+    static constexpr uint32_t IR_HARQ_RTT_MS = 4u;
     // ── CW 소거기 ON/OFF (벤치마크 비교용, 양산 기본값 true) ──
     void Set_CW_Cancel(bool enable) noexcept { cw_cancel_enabled_ = enable; }
     [[nodiscard]] bool Get_CW_Cancel() const noexcept {
